fix: move config const to constants.ts
